Add a /whitepaper.pdf symlink
Our whitepaper used to be available at /whitepaper.pdf. It is now at
/whitepaper/whitepaper.pdf. From looking at some logs, it appears that
some people are still trying to access the old URL. This change adds a
symlink pointing at the whitepaper to restore the old path. We can later
transition this to an HTTP redirect and eventually remove it completely.