Use Mailchimp API for Mailing List Subscriptions
We were previously not using the official Mailchimp API to avoid leaking the key in client-side Javascript. This key can be used to fully manage the account, so it is extremely sensitive. We are now using an AWS Lambda function with an API Gateway in front of it to allow performing subscription actions (but nothing else) while still keeping the key private.