Add new version of the newsletter signup component
This creates a new copy of the existing newsletter component, with styling differences specific to the homepage hero section. I considered refactoring the existing newsletter component so that it could work at the footer and/ or the hero section, but opted instead for the simplicity of duplicating the component because I don't expect this to be a long-lived component.