Add support for Atomic Deployments
This change re-writes the deploy script. We now deploy to a new directory within the AWS S3 bucket with a name based on the git tag/sha. The CloudFront OriginPath is then adjusted to point at the new directory. Finally, we invalidate the cache. This allows for atomic deployments and also allows us to track older versions of the site and rollback to them if necessary.